refactor: slim facade, dissolve instrumentation layer, clean re-exports
Three architectural simplifications:

1. Slim facade from 70 to 32 symbols — remove all pass-through
   re-exports from terok_sandbox and terok_agent. Consumers now import
   directly from those packages. Facade retains only terok-owned
   operations: Project/Task aggregates, task runners, task CRUD, image
   management, auth wrapper, and project state queries.

2. Dissolve lib/instrumentation/ — move agent_config.py to
   lib/domain/ where it belongs (terok-specific config stack
   composition is a domain concern). Remove the now-empty
   instrumentation layer from the architecture.

3. Clean agent_config.py re-exports — stop re-exporting
   resolve_provider_value (callers import from terok_agent directly),
   remove noqa markers from local-only imports.

Architecture layers simplified from 5 to 4:
  presentation → domain → orchestration → core

12 consumer files updated to import sandbox/agent symbols directly.
tach.toml updated: removed instrumentation layer, moved module,
slimmed facade interface.

All 1325 tests pass, module boundaries verified.
