+++ src/terok/cli/commands/gate_server.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 import argparse
 import sys
 
-from ...lib.domain.facade import (
+from terok_sandbox import (
     GateServerStatus,
     check_units_outdated,
     get_server_status,
