+++ src/terok/cli/commands/sickbay.py
@@ -23,14 +23,15 @@ import argparse
 import sys
 from pathlib import Path
 
-from ...lib.core.project_model import ProjectConfig
-from ...lib.core.projects import list_projects, load_project
-from ...lib.domain.facade import (
+from terok_sandbox import (
     check_units_outdated,
     get_container_state,
     get_server_status,
     is_systemd_available,
 )
+
+from ...lib.core.project_model import ProjectConfig
+from ...lib.core.projects import list_projects, load_project
 from ...lib.orchestration.hooks import run_hook
 from ...lib.orchestration.tasks import container_name, tasks_meta_dir
 from ...lib.util.yaml import load as _yaml_load
