+++ src/terok/lib/domain/facade.py
@@ -30,51 +30,10 @@ used by CLI commands that operate on ``project_id`` strings directly.
 from __future__ import annotations
 
 from terok_agent import (
-    AUTH_PROVIDERS,
-    AuthProvider,
     authenticate as _authenticate_raw,
 )
-from terok_sandbox import (  # noqa: F401 — re-exported public API
-    EnvironmentCheck,
-    GateServerStatus,
-    GateStalenessInfo,
-    GitGate,
-    NftNotFoundError,
-    ShieldNeedsSetup,
-    ShieldState,
-    SSHManager,
-    check_environment as shield_check_environment,
-    check_units_outdated,
-    down as shield_down,
-    get_container_state,
-    get_gate_base_path,
-    get_gate_server_port,
-    get_project_container_states,
-    get_server_status,
-    gpu_run_args,
-    install_systemd_units,
-    is_container_running,
-    is_daemon_running,
-    is_systemd_available,
-    make_shield,
-    pre_start as shield_pre_start,
-    run_setup as shield_run_setup,
-    setup_hooks_direct as shield_setup_hooks_direct,
-    start_daemon,
-    state as shield_state,
-    status as shield_status,
-    stop_daemon,
-    stop_task_containers,
-    stream_initial_logs,
-    uninstall_systemd_units,
-    up as shield_up,
-    wait_for_exit,
-)
 
-from ..core.config import (
-    SHIELD_SECURITY_HINT,  # noqa: F401 — re-exported for presentation layer
-    get_envs_base_dir,
-)
+from ..core.config import get_envs_base_dir
 from ..core.images import project_cli_image
 from ..core.projects import load_project
 from ..orchestration.docker import build_images, generate_dockerfiles
@@ -215,55 +174,14 @@ __all__ = [
     "task_logs",
     "LogViewOptions",
     # Security setup
-    "SSHManager",
-    "GitGate",
     "make_ssh_manager",
     "make_git_gate",
     # Workflow helpers
     "maybe_pause_for_ssh_key_registration",
     # Auth
-    "AUTH_PROVIDERS",
-    "AuthProvider",
     "authenticate",
-    # Gate server
-    "GateServerStatus",
-    "check_units_outdated",
-    "get_server_status",
-    "get_gate_base_path",
-    "get_gate_server_port",
-    "install_systemd_units",
-    "uninstall_systemd_units",
-    "start_daemon",
-    "stop_daemon",
-    "is_daemon_running",
-    "is_systemd_available",
-    # Git gate
-    "GateStalenessInfo",
-    "find_projects_sharing_gate",
     # Project state
     "get_project_state",
     "is_task_image_old",
-    # Container runtime (from terok_sandbox)
-    "get_container_state",
-    "get_project_container_states",
-    "gpu_run_args",
-    "is_container_running",
-    "stop_task_containers",
-    "stream_initial_logs",
-    "wait_for_exit",
-    # Shield (from terok_sandbox)
-    "SHIELD_SECURITY_HINT",
-    "EnvironmentCheck",
-    "NftNotFoundError",
-    "ShieldNeedsSetup",
-    "ShieldState",
-    "make_shield",
-    "shield_check_environment",
-    "shield_down",
-    "shield_pre_start",
-    "shield_run_setup",
-    "shield_setup_hooks_direct",
-    "shield_up",
-    "shield_state",
-    "shield_status",
+    "find_projects_sharing_gate",
 ]
