--- src/terok/lib/instrumentation/__init__.py
@@ -1,3 +1,0 @@
-# SPDX-FileCopyrightText: 2026 Jiri Vyskocil
-# SPDX-License-Identifier: Apache-2.0
-"""AI agent config, provider registry, instructions, and auth."""
