+++ src/terok/tui/widgets/task_detail.py
@@ -10,9 +10,8 @@ from rich.text import Text
 from textual.app import ComposeResult
 from textual.widgets import Static
 
-from ...lib.core.config import get_public_host
+from ...lib.core.config import SHIELD_SECURITY_HINT, get_public_host
 from ...lib.core.task_display import STATUS_DISPLAY, mode_info
-from ...lib.domain.facade import SHIELD_SECURITY_HINT
 from ...lib.orchestration.tasks import TaskMeta
 from ...lib.util.emoji import render_emoji
