+++ tach.toml
@@ -18,9 +18,8 @@ ignore_type_checking_imports = true
 # ── Layer Architecture ──────────────────────────────────────────────
 #
 #   presentation      CLI and TUI frontends
-#   domain            Project/Task aggregates, facade, task CRUD
+#   domain            Project/Task aggregates, facade, agent config, task CRUD
 #   orchestration     Container lifecycle, env/volumes, Dockerfile, ports
-#   instrumentation   Agent config, providers, instructions, wrappers
 #   core              Foundation types, config, and pure utilities
 #
 # NOTE: The sandbox layer has been extracted to the external terok-sandbox
@@ -31,7 +30,6 @@ layers = [
     "presentation",
     "domain",
     "orchestration",
-    "instrumentation",
     "core",
 ]
 
@@ -45,7 +43,7 @@ exclude = ["pydevd_pycharm", "pydevd"]
 path = "terok.cli"
 layer = "presentation"
 depends_on = [
-    "terok.lib.instrumentation.agent_config",
+    "terok.lib.domain.agent_config",
     "terok.lib.orchestration.hooks",
     "terok.lib.orchestration.tasks",
     "terok.lib.core.config",
@@ -101,7 +99,7 @@ depends_on = [
 path = "terok.lib.domain.project"
 layer = "domain"
 depends_on = [
-    "terok.lib.instrumentation.agent_config",
+    "terok.lib.domain.agent_config",
     "terok.lib.orchestration.docker",
     "terok.lib.domain.project_state",
     "terok.lib.orchestration.task_runners",
@@ -153,6 +151,13 @@ path = "terok.lib.domain.image_cleanup"
 layer = "domain"
 depends_on = ["terok.lib.core.projects"]
 
+# Agent config resolution (terok-specific stack composition)
+# Lives in domain/ but has orchestration-level layer so that task_runners can import it.
+[[modules]]
+path = "terok.lib.domain.agent_config"
+layer = "orchestration"
+depends_on = ["terok.lib.core.config", "terok.lib.core.projects"]
+
 # Project infrastructure state (queries podman & filesystem)
 [[modules]]
 path = "terok.lib.domain.project_state"
@@ -185,7 +190,7 @@ depends_on = [
 path = "terok.lib.orchestration.task_runners"
 layer = "orchestration"
 depends_on = [
-    "terok.lib.instrumentation.agent_config",
+    "terok.lib.domain.agent_config",
     "terok.lib.orchestration.container_exec",
     "terok.lib.orchestration.environment",
     "terok.lib.orchestration.hooks",
@@ -252,19 +257,6 @@ depends_on = ["terok.lib.util.yaml"]
 path = "terok.lib.orchestration.container_exec"
 layer = "orchestration"
 depends_on = ["terok.lib.util.logging_utils"]
-
-# ── INSTRUMENTATION LAYER ─────────────────────────────────────────
-#
-# NOTE: Most instrumentation modules have been extracted to the external
-# terok-agent package. Only the terok-specific config stack composition
-# (build_agent_config_stack, resolve_agent_config) remains here.
-#
-
-# Agent config resolution (terok-specific stack composition)
-[[modules]]
-path = "terok.lib.instrumentation.agent_config"
-layer = "instrumentation"
-depends_on = ["terok.lib.core.config", "terok.lib.core.projects"]
 
 # ── SANDBOX LAYER ─────────────────────────────────────────────────
 #
@@ -533,7 +525,7 @@ from = ["terok.lib.orchestration.task_runners"]
 
 [[interfaces]]
 expose = ["build_agent_config_stack", "resolve_agent_config"]
-from = ["terok.lib.instrumentation.agent_config"]
+from = ["terok.lib.domain.agent_config"]
 
 [[interfaces]]
 expose = [
@@ -655,50 +647,13 @@ expose = [
     "task_followup_headless",
     "task_logs",
     "LogViewOptions",
-    "SSHManager",
-    "GitGate",
-    "AUTH_PROVIDERS",
-    "AuthProvider",
-    "authenticate",
-    "GateServerStatus",
-    "check_units_outdated",
-    "get_server_status",
-    "get_gate_base_path",
-    "get_gate_server_port",
-    "install_systemd_units",
-    "uninstall_systemd_units",
-    "start_daemon",
-    "stop_daemon",
-    "is_daemon_running",
-    "is_systemd_available",
-    "GateStalenessInfo",
     "find_projects_sharing_gate",
     "make_git_gate",
     "make_ssh_manager",
+    "authenticate",
     "get_project_state",
     "is_task_image_old",
-    "make_shield",
     "maybe_pause_for_ssh_key_registration",
-    "EnvironmentCheck",
-    "NftNotFoundError",
-    "ShieldNeedsSetup",
-    "ShieldState",
-    "shield_check_environment",
-    "shield_down",
-    "shield_run_setup",
-    "shield_setup_hooks_direct",
-    "shield_pre_start",
-    "shield_up",
-    "shield_state",
-    "shield_status",
-    "SHIELD_SECURITY_HINT",
-    "get_container_state",
-    "get_project_container_states",
-    "gpu_run_args",
-    "is_container_running",
-    "stop_task_containers",
-    "stream_initial_logs",
-    "wait_for_exit",
 ]
 from = ["terok.lib.domain.facade"]
